Add files via upload
At ALX, one of our first tasks is to write a pseudocode to calculate the sum of squares of given input of numbers in a google doc and give access to anyone with the link. The document has been uploaded to help others who have no idea what this means as I had some struggle with it because I am a beginning beginner with no prior tech knowledge.